refactor(components): Arrange new components folder.
- Move components to a new components folder and fix imports.
- Update repo comment component to bind data from dialog parent component.
- Update dialog method to not pass html as argument.
