+++ src/app/app.module.ts
@@ -4,12 +4,12 @@ import { NgModule } from '@angular/core';
 import { MatDialogModule } from '@angular/material';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { GitSearchComponent } from './git-search/git-search.component';
+import { GitSearchComponent } from './components/git-search/git-search.component';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule } from '@angular/forms';
-import { RepoComponent } from './repo/repo.component';
-import { DialogComponent } from './dialog/dialog.component';
-import { RepoCommentComponent } from './repo-comment/repo-comment.component';
+import { RepoComponent } from './components/repo/repo.component';
+import { DialogComponent } from './components/dialog/dialog.component';
+import { RepoCommentComponent } from './components/repo-comment/repo-comment.component';
 
 
 @NgModule({
