+++ src/app/components/repo-comment/repo-comment.component.html
@@ -1,0 +1,1 @@
+<p><strong>{{comment.user.login}}</strong> - <i>"{{comment.body}}"</i></p>
