--- src/app/dialog/dialog.component.html
@@ -1,6 +1,0 @@
-<h1 mat-dialog-title>Comments</h1>
-<mat-dialog-content>
-  <!-- {{data.content}} -->
-  <div [innerHtml]="data.content">
-  </div>
-</mat-dialog-content>
