--- src/app/dialog/dialog.component.ts
@@ -1,11 +1,0 @@
-import { Component, ContentChild, Inject} from '@angular/core';
-import { MAT_DIALOG_DATA } from '@angular/material';
-
-@Component({
-  selector: 'app-dialog',
-  templateUrl: './dialog.component.html'
-})
-export class DialogComponent {
-  constructor(@Inject(MAT_DIALOG_DATA) public data: any ) {
-  }
-}
