--- src/app/git-search/git-search.component.html
@@ -1,21 +1,0 @@
-<div style="text-align:center">
-  <h1>{{title}}</h1>
-  <input type="text" name="searchInput" placeholder="Enter the name of the repo" [(ngModel)]="model.q" required appNospecialChars>
-  <!-- <input [type]="text" name="searchInput" placeholder="Enter the name of the repo" required appNospecialChars> -->
-
-  <!-- <p class="error" *ngIf="query.value === ''">Query value is required!</p>
-  <p class="error" *ngIf="query.dirty && query.value !== '' && query.invalid">You need to type a valid query without special characters</p>
-  <p>{{query.name}} value '{{query.value}}' is {{isNumber(query.value) ? 'number' : 'text'}}</p> -->
-  
-  <form #searchForm="ngForm" (ngSubmit)="sendQuery()">
-    <button type="submit" id="searchButton" [disabled]=!searchForm.valid>Submit</button>
-  </form>
-</div>
-<div *ngIf="searchResults; else noResults">
-  <!-- <h3>Results: {{searchResults.total_count}}</h3> -->
-  <h3>Results: {{searchResults.total_count}}</h3>
-  <div *ngFor="let repoItem of searchResults.items">
-    <app-repo [keys]="repoItem" (click)="openDialog(repoItem.comments_url)"></app-repo>
-  </div>
-</div>
-<ng-template #noResults>Loading...</ng-template>
