--- src/app/repo-comment/repo-comment.component.html
@@ -1,6 +1,0 @@
-<p *ngIf="!empty; else emptyComments">
-  {{keys.user}} <i>"{{keys.body}}"</i>
-</p>
-<p #emptyComments>
-  Repository seems to have no comments yet.
-</p>
